pyproject.toml
[tool.poetry]
name = "image-app"
version = "0.1.0"
description = ""
authors = ["Rio M. <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
numpy = "^1.18.4"
sqlalchemy = "^1.3.16"
Pillow = "^7.1.2"
psycopg2-binary = "^2.8.5"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
coverage = "^5.1"
tensorflow = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
